Rewired Left Hand Side and Edge Cuts
Rewired the entire left hand side of the keyboard. Also added the edge cuts for the outer shape of each board. Redid this part on the left hand side to make it match the design of the right hand side.
Will start working on the rewiring of the right hand side next.